feat: use playwright, adapt account tests
